style(suite): refactor file names for Buy/Sell quoted, remove default exports
