+++ packages/suite/src/views/wallet/coinmarket/buy/offers/Offers/List/BuyQuoteList.tsx
@@ -1,19 +1,20 @@
 import styled from 'styled-components';
-import { SellFiatTrade } from 'invity-api';
-
 import { CoinLogo, variables, Icon, H2 } from '@trezor/components';
-import { useCoinmarketSellOffersContext } from 'src/hooks/wallet/useCoinmarketSellOffers';
+import { BuyTrade } from 'invity-api';
+import { useCoinmarketBuyOffersContext } from 'src/hooks/wallet/useCoinmarketBuyOffers';
 import { Translation } from 'src/components/suite';
-import { CoinmarketRefreshTime } from 'src/views/wallet/coinmarket/common';
+import {
+    CoinmarketCryptoAmount,
+    CoinmarketFiatAmount,
+    CoinmarketRefreshTime,
+} from 'src/views/wallet/coinmarket/common';
 import { InvityAPIReloadQuotesAfterSeconds } from 'src/constants/wallet/coinmarket/metadata';
-import { CoinmarketCryptoAmount } from 'src/views/wallet/coinmarket/common/CoinmarketCryptoAmount';
-import { CoinmarketFiatAmount } from 'src/views/wallet/coinmarket/common/CoinmarketFiatAmount';
-import Quote from './Quote';
+import { BuyQuote } from './BuyQuote';
 
 const Wrapper = styled.div``;
 const Quotes = styled.div``;
 
-const StyledQuote = styled(Quote)`
+const StyledBuyQuote = styled(BuyQuote)`
     margin-bottom: 20px;
 `;
 
@@ -40,10 +41,9 @@ const Right = styled.div`
     }
 `;
 
-const SummaryRow = styled(H2)`
+const SummaryRow = styled.div`
     display: flex;
     align-items: center;
-    font-weight: ${variables.FONT_WEIGHT.REGULAR};
 `;
 
 const OrigAmount = styled.div`
@@ -58,13 +58,12 @@ const StyledIcon = styled(Icon)`
 
 const Send = styled(H2)`
     padding-top: 3px;
-    padding-left: 10px;
     font-weight: ${variables.FONT_WEIGHT.REGULAR};
 `;
 
 const Receive = styled(H2)`
     padding-top: 3px;
-    padding-right: 10px;
+    padding-left: 10px;
     font-weight: ${variables.FONT_WEIGHT.REGULAR};
 `;
 
@@ -78,36 +77,36 @@ const NoQuotes = styled.div`
 
 interface ListProps {
     isAlternative?: boolean;
-    quotes: SellFiatTrade[];
+    quotes: BuyTrade[];
 }
 
-const List = ({ isAlternative, quotes }: ListProps) => {
-    const { account, quotesRequest, timer } = useCoinmarketSellOffersContext();
+export const BuyQuoteList = ({ isAlternative, quotes }: ListProps) => {
+    const { account, quotesRequest, timer } = useCoinmarketBuyOffersContext();
 
     if (!quotesRequest) return null;
-    const { fiatStringAmount, fiatCurrency, amountInCrypto } = quotesRequest;
+    const { fiatStringAmount, fiatCurrency, wantCrypto } = quotesRequest;
 
     return (
-        <Wrapper>
+        <Wrapper data-test="@coinmarket/buy/offers-list">
             <Header>
                 <Left>
                     <SummaryRow>
-                        <CoinLogo size={21} symbol={account.symbol} />
                         <Send>
-                            <CoinmarketCryptoAmount
-                                amount={amountInCrypto ? quotes[0].cryptoStringAmount : ''}
-                                symbol={account.symbol}
+                            <CoinmarketFiatAmount
+                                amount={!wantCrypto ? quotes[0].fiatStringAmount : ''}
+                                currency={quotes[0].fiatCurrency}
                             />
                         </Send>
                         <StyledIcon icon="ARROW_RIGHT_LONG" />
+                        <CoinLogo size={21} symbol={account.symbol} />
                         <Receive>
-                            <CoinmarketFiatAmount
-                                amount={!amountInCrypto ? quotes[0].fiatStringAmount : ''}
-                                currency={quotes[0].fiatCurrency}
+                            <CoinmarketCryptoAmount
+                                amount={wantCrypto ? quotes[0].receiveStringAmount : ''}
+                                symbol={account.symbol}
                             />
                         </Receive>
                     </SummaryRow>
-                    {isAlternative && !amountInCrypto && (
+                    {isAlternative && !wantCrypto && (
                         <OrigAmount>
                             ≈{' '}
                             <CoinmarketFiatAmount
@@ -123,7 +122,7 @@ const List = ({ isAlternative, quotes }: ListProps) => {
                             isLoading={timer.isLoading}
                             refetchInterval={InvityAPIReloadQuotesAfterSeconds}
                             seconds={timer.timeSpend.seconds}
-                            label={<Translation id="TR_SELL_OFFERS_REFRESH" />}
+                            label={<Translation id="TR_BUY_OFFERS_REFRESH" />}
                         />
                     </Right>
                 )}
@@ -131,13 +130,13 @@ const List = ({ isAlternative, quotes }: ListProps) => {
             <Quotes>
                 {quotes?.length === 0 ? (
                     <NoQuotes>
-                        <Translation id="TR_SELL_NO_OFFERS" />
+                        <Translation id="TR_BUY_NO_OFFERS" />
                     </NoQuotes>
                 ) : (
                     quotes.map(quote => (
-                        <StyledQuote
-                            amountInCrypto={amountInCrypto}
-                            key={`${quote.exchange}-${quote.paymentMethod}-${quote.fiatCurrency}`}
+                        <StyledBuyQuote
+                            wantCrypto={wantCrypto}
+                            key={`${quote.exchange}-${quote.paymentMethod}-${quote.receiveCurrency}`}
                             quote={quote}
                         />
                     ))
@@ -146,5 +145,3 @@ const List = ({ isAlternative, quotes }: ListProps) => {
         </Wrapper>
     );
 };
-
-export default List;
