+++ packages/suite/src/views/wallet/coinmarket/sell/offers/Offers/index.tsx
@@ -10,8 +10,8 @@ import { Translation } from 'src/components/suite';
 import { useLayout } from 'src/hooks/suite/useLayout';
 import { useCoinmarketSellOffersContext } from 'src/hooks/wallet/useCoinmarketSellOffers';
 import { useCoinmarketNavigation } from 'src/hooks/wallet/useCoinmarketNavigation';
-import List from './List';
 import { SelectedOffer } from './SelectedOffer';
+import { SellQuoteList } from './List/SellQuoteList';
 
 const Wrapper = styled.div`
     padding: 16px 32px 32px;
@@ -84,7 +84,7 @@ const Offers = () => {
                         />
                     ) : (
                         <>
-                            <List quotes={quotes} />
+                            <SellQuoteList quotes={quotes} />
                             {alternativeQuotes.length > 0 && (
                                 <>
                                     <Divider>
@@ -95,7 +95,7 @@ const Offers = () => {
                                         </DividerMiddle>
                                         <DividerRight />
                                     </Divider>
-                                    <List isAlternative quotes={alternativeQuotes} />
+                                    <SellQuoteList isAlternative quotes={alternativeQuotes} />
                                 </>
                             )}
                         </>
